// src/costs/constants/paths.enum.ts
export enum Paths {
    COSTS = "costs",
    PARAM_COST = ":costId",
    PARAM_COST_ID = "costId",

    ITEM = "item",
    PARAM_ITEM = ":itemId",
    PARAM_ITEM_ID = "itemId",

    CONSTANT = "constant",
    PARAM_CONSTANT = ':constantId',
    PARAM_CONSTANT_ID = "constantId",

    CALCULATE = "calculate",

    RULE = "rule",
    PARAM_RULE = ":ruleId",
    PARAM_RULE_ID = "ruleId"

}